requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
google-generativeai==0.7.2
faiss-cpu==1.8.0
pydantic==2.8.2
python-dotenv==1.0.1
tiktoken==0.7.0
